Cleanup enqueue and stats API
* Use mixing for EnqueueApi
* Remove unused API calls
